requirements.txt
discord-webhook
ping3==4.0.4
huawei-lte-api
requests
python-dotenv